Removed index.js from actions and moved actions to their respective associated files
